// src/app/model/representants.model.ts
export class Representant{
    idRepresentant: number;
    prenomRepresentant: string;
    nomRepresentant: string;
    civiliteRepresentant: string;
    fonctionRepresentant: string;
    emailRepresentant: string;
    telephoneFixeRepresentant: number;
    posteRepresentant: number;
    telephoneMobileRepresentant: number;
    habilitationRepresentant: string;
   // pointFocalRepresentant: boolean;
}